404 page
Wanted to create a function, that would redirect the user to a 404 page if the user tried to navigate the page using the url. for example if he tried to go to the dashboard he could simply write localhost:300/dashboard.html. but now the user cant. it still possible with other pages though.